Finished implementing goroutines style syntax

+++ crates/std/src/stdlib/thread/main.cl
@@ -1,6 +1,6 @@
-extern "c" const c_usleep = fn(@uint) -> @int from "libc" as "usleep";
+extern "c" const c_usleep = fn(@int) -> @int from "libc" as "usleep";
 
 const wait = fn(ms: int) => if ms <= 0 => null else => {
-    c_usleep((ms as uint) * 1000);
+    c_usleep(ms * 1000);
     null;
 };
